Model/requirements.txt
Flask==2.2.3
pymongo==4.3.3
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.1.3
python-dotenv==0.21.0
bson==0.5.10
logging==0.5.1.2
